requirements.txt
PyPDF2
langchain-text-splitters
transformers
torch
faiss-cpu